Move JSON and Slicer CLI task endpoint logic to separate files
