+++ plugins/item_tasks/server/rest.py
@@ -4,11 +4,13 @@ import six
 from girder import events
 from girder.api import access
 from girder.api.describe import autoDescribeRoute, Description
-from girder.api.rest import ensureTokenScopes, filtermodel, Resource, RestException
+from girder.api.rest import ensureTokenScopes, filtermodel, Resource
 from girder.constants import AccessType, TokenScope
 from girder.models.model_base import ValidationException
 from girder.plugins.worker import utils
-from . import cli_parser, constants
+from . import constants
+from .json_tasks import runJsonTasksDescription, addJsonSpecs
+from .slicer_cli_tasks import runSlicerCliDescription, setSpecFromXml
 
 
 class ItemTask(Resource):
@@ -19,10 +21,10 @@ class ItemTask(Resource):
 
         self.route('GET', (), self.listTasks)
         self.route('POST', (':id', 'execution'), self.executeTask)
-        self.route('POST', (':id', 'slicer_cli_description'), self.runSlicerCliDescription)
-        self.route('PUT', (':id', 'slicer_cli_xml'), self.setSpecFromXml)
-        self.route('POST', (':id', 'json_description'), self.runJsonTasksDescription)
-        self.route('POST', (':id', 'json_specs'), self.addJsonSpecs)
+        self.route('POST', (':id', 'slicer_cli_description'), runSlicerCliDescription)
+        self.route('PUT', (':id', 'slicer_cli_xml'), setSpecFromXml)
+        self.route('POST', (':id', 'json_description'), runJsonTasksDescription)
+        self.route('POST', (':id', 'json_specs'), addJsonSpecs)
 
     @access.public
     @autoDescribeRoute(
@@ -201,231 +203,3 @@ class ItemTask(Resource):
         jobModel.scheduleJob(job)
 
         return job
-
-    @access.admin(scope=constants.TOKEN_SCOPE_AUTO_CREATE_CLI)
-    @filtermodel(model='job', plugin='jobs')
-    @autoDescribeRoute(
-        Description('Create an item task spec based on a docker image.')
-        .notes('This operates on an existing item, turning it into an item task '
-               'using Slicer CLI introspection of a docker image.')
-        .modelParam('id', 'The ID of the item that the task spec will be bound to.',
-                    model='item', level=AccessType.WRITE)
-        .param('image', 'The docker image name. If not passed, uses the existing'
-               'itemTaskSpec.docker_image metadata value.', required=False, strip=True)
-        .jsonParam('args', 'Arguments to be passed to the docker container to output the '
-                   'Slicer CLI spec.', required=False, default=[], requireArray=True)
-        .param('setName', 'Whether item name should be changed to the title of the CLI.',
-               dataType='boolean', required=False, default=True)
-        .param('setDescription', 'Whether the item description should be changed to the '
-               'description of the CLI.', dataType='boolean', required=False, default=True)
-        .param('pullImage', 'Whether the image should be pulled from a docker registry. '
-               'Set to false to use local images only.', dataType='boolean', required=False,
-               default=True)
-    )
-    def runSlicerCliDescription(
-            self, item, image, args, setName, setDescription, pullImage, params):
-        if 'meta' not in item:
-            item['meta'] = {}
-
-        if image is None:
-            image = item.get('meta', {}).get('itemTaskSpec', {}).get('docker_image')
-
-        if not image:
-            raise RestException(
-                'You must pass an image parameter, or set the itemTaskSpec.docker_image '
-                'field of the item.')
-
-        jobModel = self.model('job', 'jobs')
-        token = self.model('token').createToken(
-            days=3, scope='item_task.set_task_spec.%s' % item['_id'])
-        job = jobModel.createJob(
-            title='Read docker Slicer CLI: %s' % image, type='item_task.slicer_cli',
-            handler='worker_handler', user=self.getCurrentUser())
-
-        if args[-1:] == ['--xml']:
-            args = args[:-1]
-
-        job.update({
-            'itemTaskId': item['_id'],
-            'kwargs': {
-                'task': {
-                    'mode': 'docker',
-                    'docker_image': image,
-                    'container_args': args + ['--xml'],
-                    'pull_image': pullImage,
-                    'outputs': [{
-                        'id': '_stdout',
-                        'format': 'text'
-                    }],
-                },
-                'outputs': {
-                    '_stdout': {
-                        'mode': 'http',
-                        'method': 'PUT',
-                        'format': 'text',
-                        'url': '/'.join((utils.getWorkerApiUrl(), self.resourceName,
-                                         str(item['_id']), 'slicer_cli_xml')),
-                        'params': {
-                            'setName': setName,
-                            'setDescription': setDescription
-                        },
-                        'headers': {'Girder-Token': token['_id']}
-                    }
-                },
-                'jobInfo': utils.jobInfoSpec(job),
-                'validate': False,
-                'auto_convert': False,
-                'cleanup': True
-            }
-        })
-
-        item['meta']['itemTaskSpec'] = {
-            'mode': 'docker',
-            'docker_image': image
-        }
-
-        if args:
-            item['meta']['itemTaskSlicerCliArgs'] = args
-
-        self.model('item').save(item)
-
-        job = jobModel.save(job)
-        jobModel.scheduleJob(job)
-
-        return job
-
-    @access.token
-    @autoDescribeRoute(
-        Description('Set a task spec on an item from a Slicer CLI XML spec.')
-        .modelParam('id', model='item', force=True)
-        .param('xml', 'The Slicer CLI XML spec.', paramType='body')
-        .param('setName', 'Whether item name should be changed to the title of the CLI.',
-               dataType='boolean', required=True)
-        .param('setDescription', 'Whether the item description should be changed to the '
-               'description of the CLI.', dataType='boolean', required=False, default=True),
-        hide=True
-    )
-    def setSpecFromXml(self, item, xml, setName, setDescription, params):
-        self.ensureTokenScopes('item_task.set_task_spec.%s' % item['_id'])
-
-        args = item.get('meta', {}).get('itemTaskSlicerCliArgs') or []
-        cliSpec = cli_parser.parseSlicerCliXml(xml)
-
-        itemTaskSpec = item.get('meta', {}).get('itemTaskSpec', {})
-        itemTaskSpec.update({
-            'container_args': args + cliSpec['args'],
-            'inputs': cliSpec['inputs'],
-            'outputs': cliSpec['outputs']
-        })
-
-        if setName:
-            item['name'] = cliSpec['title']
-        if setDescription:
-            item['description'] = cliSpec['description']
-
-        self.model('item').setMetadata(item, {
-            'itemTaskSpec': itemTaskSpec,
-            'isItemTask': True
-        })
-
-    @access.admin(scope=constants.TOKEN_SCOPE_AUTO_CREATE_CLI)
-    @filtermodel(model='job', plugin='jobs')
-    @autoDescribeRoute(
-        Description('Create item task specs based on a docker image.')
-        .notes('This operates on an existing folder, adding item tasks '
-               'using introspection of a docker image.')
-        .modelParam('id', 'The ID of the folder that the task specs will be added to.',
-                    model='folder', level=AccessType.WRITE)
-        .param('image', 'The docker image name.', required=True, strip=True)
-        .param('pullImage', 'Whether the image should be pulled from Docker Hub. '
-               'Set to false to use local images only.',
-               dataType='boolean', required=False, default=True)
-    )
-    def runJsonTasksDescription(self, folder, image, pullImage, params):
-        jobModel = self.model('job', 'jobs')
-        token = self.model('token').createToken(
-            days=3, scope='item_task.set_task_spec.%s' % folder['_id'],
-            user=self.getCurrentUser())
-        job = jobModel.createJob(
-            title='Read docker task specs: %s' % image, type='item_task.json_description',
-            handler='worker_handler', user=self.getCurrentUser())
-
-        jobOptions = {
-            'itemTaskId': folder['_id'],
-            'kwargs': {
-                'task': {
-                    'mode': 'docker',
-                    'docker_image': image,
-                    'container_args': [],
-                    'pull_image': pullImage,
-                    'outputs': [{
-                        'id': '_stdout',
-                        'format': 'text'
-                    }],
-                },
-                'outputs': {
-                    '_stdout': {
-                        'mode': 'http',
-                        'method': 'POST',
-                        'format': 'text',
-                        'url': '/'.join((utils.getWorkerApiUrl(), self.resourceName,
-                                         str(folder['_id']), 'json_specs')),
-                        'headers': {'Girder-Token': token['_id']},
-                        'params': {
-                            'image': image,
-                            'pullImage': pullImage
-                        }
-                    }
-                },
-                'jobInfo': utils.jobInfoSpec(job),
-                'validate': False,
-                'auto_convert': False,
-                'cleanup': True
-            }
-        }
-        job.update(jobOptions)
-
-        job = jobModel.save(job)
-        jobModel.scheduleJob(job)
-        return job
-
-    @access.token
-    @autoDescribeRoute(
-        Description('Create item tasks under a folder using a list of JSON specifications.')
-        .modelParam('id', model='folder', force=True)
-        .jsonParam('json', 'The JSON specifications as a list or a single object.',
-                   paramType='body')
-        .param('image', 'The docker image name.', required=True, strip=True)
-        .param('pullImage', 'Whether the image should be pulled from Docker Hub. ' +
-               'Set to false to use local images only.',
-               dataType='boolean', required=False, default=True),
-        hide=True
-    )
-    def addJsonSpecs(self, folder, json, image, pullImage, params):
-        self.ensureTokenScopes('item_task.set_task_spec.%s' % folder['_id'])
-        token = self.getCurrentToken()
-        user = self.model('user').load(token['userId'], force=True)
-
-        if not isinstance(json, list):
-            json = [json]
-
-        for itemIndex, itemTaskSpec in enumerate(json):
-            name = itemTaskSpec.get('name')
-            if not name:
-                name = image
-                if len(json) > 1:
-                    name += ' ' + str(itemIndex)
-
-            item = self.model('item').createItem(
-                name=name,
-                creator=user,
-                folder=folder,
-                description=itemTaskSpec.get('description', ''),
-                reuseExisting=True)
-
-            itemTaskSpec['docker_image'] = image
-            itemTaskSpec['pull_image'] = pullImage
-            self.model('item').setMetadata(item, {
-                'itemTaskSpec': itemTaskSpec,
-                'isItemTask': True
-            })
